Corrected a bug in Mars_Year calculation
The original code was fortuitously correct for year 0, but increases for dates
below the start of year 0, instead of decreasing. This was because the calculation for
the differences were the wrong way around. This correction also means that an offset factor should have been included in the calculation (as it was for the math component) because of the use of the floor function.

Incremented the number and added a new test.